+++ packages/browser/src/browser/standalone.ts
@@ -2,10 +2,10 @@
 import { getCDN, setGlobalCDNUrl } from '../lib/parse-cdn'
 import { setVersionType } from '../lib/version-type'
 
-if (process.env.ASSET_PATH) {
-  if (process.env.ASSET_PATH === '/dist/umd/') {
+if (process.env.IS_WEBPACK_BUILD) {
+  if (process.env.ASSET_PATH) {
     // @ts-ignore
-    __webpack_public_path__ = '/dist/umd/'
+    __webpack_public_path__ = process.env.ASSET_PATH
   } else {
     const cdn = getCDN()
     setGlobalCDNUrl(cdn)