+++ packages/consent/consent-wrapper-onetrust/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @segment/analytics-consent-wrapper-onetrust
 
+## 1.1.1
+
+### Patch Changes
+
+- [#1136](https://github.com/segmentio/analytics-next/pull/1136) [`a648fdb`](https://github.com/segmentio/analytics-next/commit/a648fdbfa8efec04f5f0ba5205639e2c3a442068) Thanks [@silesky](https://github.com/silesky)! - Fix consent namespace issue
+
+## 1.1.0
+
+### Minor Changes
+
+- [#1133](https://github.com/segmentio/analytics-next/pull/1133) [`3abc609`](https://github.com/segmentio/analytics-next/commit/3abc6090c3bf1c7cf4947c181555c5893bc8626d) Thanks [@silesky](https://github.com/silesky)! - Add RequireJS support (analytics-onetrust.js)
+
 ## 1.0.1
 
 ### Patch Changes